Updated types of offer

+++ src/main/java/app/preciojusto/products/entities/OfferUnitPercentage.java
@@ -14,5 +14,5 @@ public class OfferUnitPercentage extends Offer {
 
     private Integer ofupunitaffected;
 
-    private Integer ofuppercentage;
+    private Double ofuppercentage;
 }
